requirements.txt
# Requirements
libusb1
numpy
opencv-python
